Typing of the variables
refs #4596

+++ oscar-linprog/src/main/examples/oscar/examples/linprog/ElectricityMarket.scala
@@ -19,6 +19,7 @@ import oscar.algebra._
 import oscar.linprog.interface.MIPSolverInterface
 import oscar.linprog.interface.lpsolve.LPSolveLib
 import oscar.linprog.modeling._
+import Migration._
 
 import scala.io.Source
 
@@ -60,7 +61,7 @@ object ElectricityMarket extends MPModel(LPSolveLib) with App {
     val cons_t = consumers.filter(overlap(_, t))
     if (prod_t.nonEmpty && cons_t.nonEmpty) {
       // production and consumption must be the same at time t
-      add( s"C_${solver.getNumberOfLinearConstraints}" ||: sum(prod_t)(p => varMap(p) * p(0).abs) === sum(cons_t)(c => varMap(c) * c(0).abs))
+      add( s"C_${solver.getNumberOfLinearConstraints}" ||: sum(prod_t)(p => varMap(p) * p(0).abs.toDouble) === sum(cons_t)(c => varMap(c) * c(0).abs.toDouble))
     }
   }
